+++ changes.md
@@ -1,4 +1,6 @@
 #### 0.3.1
+* Added `insert_media` API/RPC endpoints and improved `insert_media` utility method to wipe all old attachments.
+* Allowing manual override of `ud_site_id` and `ud_site_secret_token` via constants.
 * Improved support for invoking `get_schedule_listings` via wp-rets or xmlrpc.
 * Added option to `skipTermCounting` when doing `wpp.editProperty`
 * Changed `wpp.scheduleStats` to returns counts of only published, private and future listings.